renamed lgTelemtryUtil to telemetryUtils
