+++ libraries/botbuilder-dialogs-adaptive/tests/telemetryUtils.js
@@ -7,13 +7,13 @@ const { NullTelemetryClient } = require('botbuilder-core');
  */
 function createTelemetryClientAndStub(captureTelemetryAction) {
     if (captureTelemetryAction && typeof captureTelemetryAction !== 'function') {
-        throw new TypeError(`Failed test arrangement - createtelemetryClientAndStub() received ${typeof captureTelemetryAction} instead of undefined or a function.`);
+        throw new TypeError(`Failed test arrangement - createtelemetryClientAndStub() received ${ typeof captureTelemetryAction } instead of undefined or a function.`);
     }
 
     function wrapAction(...args) {
         captureTelemetryAction(...args);
     }
-    
+
     const telemetryClient = new NullTelemetryClient();
     const trackEventStub = stub(telemetryClient, 'trackEvent');
